refactor(render): mechanical split of render/discopy/translator.py into sibling modules
MAJ-04 module 4/6: render/discopy/translator.py 2150 -> 303 lines (facade).

- bootstrap.py: DisCoPy/JAX globals, availability probing, init, setup reports
- gnn_parsing.py: GNN content parsing helpers (leaf)
- diagram_builders.py: categorical diagram builders
- matrix_builders.py: matrix-backed diagram builders
- file_translation.py: file-level translation orchestrators
- code_templates.py: standalone script template emitters
- translator.py: facade re-exporting all 36 moved names; the standalone
  __main__ smoke harness retained verbatim

DAG: bootstrap <- {gnn_parsing, diagram_builders, matrix_builders} <- file_translation; code_templates independent; facade on top. 2108 lines verified byte-identical against the pre-split file; only import headers differ. Single-owner mutable globals (the 15 component slots and 5 availability flags live only in bootstrap; import-binding snapshots are safe because nothing re-calls initialize_discopy_components() after import). Package __init__ and all consumers unchanged.

Verified: ruff 0 violations (repo-wide check); mypy 0 errors; facade contract 36 names with object identity (Box is Box); targeted pytest test_discopy_improvements + test_render_overall: 12 passed.
